requirements.txt
annotated-types==0.7.0
anyio==4.7.0
cffi==1.17.1
click==8.1.7
cryptography==44.0.0
ecdsa==0.19.0
fastapi==0.114.0
greenlet==3.1.1
h11==0.14.0
idna==3.10
pyasn1==0.6.1
pycparser==2.22
pydantic==2.9.0
pydantic_core==2.23.2
python-jose==3.3.0
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.34
starlette==0.38.6
typing_extensions==4.12.2
tzdata==2024.2
uvicorn==0.30.6
